Add tests for signature views.
